Add files via upload
Humans Deserve the Truth. In today’s world, mobile phones are an invaluable asset, containing both personal and professional information. Losing such a device can have serious repercussions, both financially and emotionally. This system was initiated because of personal experiences of losing a mobile phone.  
 
In order to address this issue, we have developed PhoneLoc, a software solution designed to retrieve the location of a misplaced phone. This powerful code uses python packages to obtain confidential information about the phone number entered, such as the country of origin and the name of the current service provider. Furthermore, it takes the coordinates of the current location of the phone Number and extracts the latitude and longitude location of the device.